% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FGSPCAUtils.R
\name{convcheck}
\alias{convcheck}
\title{Convergence check of two successive matrices}
\usage{
convcheck(beta1, beta2)
}
\arguments{
\item{beta1}{\eqn{B_1} a matrix}

\item{beta2}{\eqn{B_2} a matrix}
}
\value{
the largest absolute value of the difference of their corresponding columns
}
\description{
Check the convergence of two successive matrices
}
\details{
The two matrices \eqn{B_1} and \eqn{B_2} may be different only in their column signs
with respective to each column.
}
