% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FGSPCAUtils.R
\name{coordinate_descent_enet}
\alias{coordinate_descent_enet}
\title{The coordinate descent for the standard elastic net regression problem}
\usage{
coordinate_descent_enet(
  x,
  y,
  lambda,
  alpha,
  max.steps = 100,
  condition_tol = 0.001,
  loss_return = FALSE
)
}
\arguments{
\item{x}{the data matrix \eqn{X_{n\times p}}}

\item{y}{the response vector \eqn{Y_{n \times 1}}}

\item{lambda}{the \eqn{\lambda} in the elastic net loss function}

\item{alpha}{the \eqn{\alpha} in the elastic net loss function}

\item{max.steps}{maximum steps, the maximum number of steps for the updating, \code{100} (default)}

\item{condition_tol}{the tolerance for the condition to stop, \code{1e-3} (default)}

\item{loss_return}{whether to return loss or not, \code{FALSE} (default)}
}
\value{
\eqn{\beta} the solution \eqn{\beta} or a list

a list of \eqn{\beta}, a sequence of the objective values during the CD process,
a sequence of the loss values,  a sequence of the penalty values,
\code{list(beta=beta, obj=obj, loss=loss, pen=pen)}.
}
\description{
The objective function of the standard elastic net regression problem is
\deqn{1/(2n) \| Y-X\beta\|_2^2 + \lambda (\alpha \|\beta\|_1 + (1-\alpha )/2\|\beta\|_2^2) .}
It is the standard version of the elastic net problem.
}
