% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FGSPCAUtils.R
\name{coordinate_descent_LASSO}
\alias{coordinate_descent_LASSO}
\title{The coordinate descent for LASSO (elastic net) regression}
\usage{
coordinate_descent_LASSO(
  x,
  y,
  paras,
  max.steps = 100,
  condition_tol = 0.001,
  loss_return = FALSE
)
}
\arguments{
\item{x}{the data matrix \eqn{X_{n\times p}}}

\item{y}{the response vector \eqn{Y_{n \times 1}}}

\item{paras}{the combination of parameters \eqn{\lambda_2, \lambda_1}}

\item{max.steps}{maximum steps, the maximum number of steps for the updating, \code{100} (default)}

\item{condition_tol}{the tolerance for the condition to stop, \code{1e-3} (default)}

\item{loss_return}{whether to return loss or not, \code{FALSE} (default)}
}
\value{
\eqn{\beta} the solution \eqn{\beta} or a list

a list of \eqn{\beta}, a sequence of the objective values during the CD process,
a sequence of the loss values,  a sequence of the penalty values,
\code{list(beta=beta, obj=obj, loss=loss, pen=pen)}.
}
\description{
The coordinate descent to solve the LASSO (elastic net) regression problem.
It is not the standard version of the elastic net problem, so we use \code{coordinate_descent_LASSO}.
The standard version of the elastic net problem is termed \code{coordinate_descent_enet}.
}
\details{
The objective function of the lasso (elastic net) is
\deqn{1/(2n) \| Y-X\beta\|_2^2 + \lambda_1\|\beta\|_1 + \lambda_2\|\beta\|_2^2 .}
While the soft thresholding method applied to the matrix version may not lead to convergence,
the updating using the soft thresholding along each coordinate can guarantee the convergence.
}
