% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FGSPCAUtils.R
\name{different_beta_solver}
\alias{different_beta_solver}
\title{Different beta solvers}
\usage{
different_beta_solver(
  x,
  y,
  lambda2,
  lambda1,
  sparse,
  solver_type = "spcasolver"
)
}
\arguments{
\item{x}{the data matrix \eqn{X_{n\times p}}}

\item{y}{the response vector \eqn{Y_{n\times 1}}}

\item{lambda2}{\eqn{\lambda_2} in the elastic loss function}

\item{lambda1}{\eqn{\lambda_1} in the elastic loss function}

\item{sparse}{\code{sparse = c("penalty", "varnum")}}

\item{solver_type}{\code{solver_type=c("spcasolver", "lassocd", "glmnet")}}
}
\value{
the solution \eqn{\beta}
}
\description{
Different beta solvers
}
