% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FGSPCAUtils.R
\name{enet_loss}
\alias{enet_loss}
\title{The elastic net loss function}
\usage{
enet_loss(x, y, beta, lambda1, lambda2, verbose = FALSE)
}
\arguments{
\item{x}{the data matrix \eqn{X_{n\times p}}}

\item{y}{the response vector \eqn{Y_{n \times 1}}}

\item{beta}{\eqn{\beta} the estimation of \eqn{\beta}}

\item{lambda1}{the \eqn{\lambda_1} in the loss function}

\item{lambda2}{the \eqn{\lambda_2} in the loss function}

\item{verbose}{whether to return a list, \code{FALSE} (default)}
}
\value{
the value of the loss function or a list \code{c(loss, pen, obj)}
}
\description{
The elastic net loss function (not the standard version)
}
\details{
The standard version of the elastic net problem is defined as follows,
\deqn{1/(2n) \| Y-X\beta\|_2^2 + \lambda (\alpha \|\beta\|_1 + (1-\alpha )/2\|\beta\|_2^2) .}

The objective function of the elastic net we use in this function is defined as
\deqn{1/(2n) \| Y-X\beta\|_2^2 + \lambda_1\|\beta\|_1 + \lambda_2\|\beta\|_2^2 .}
It is not the standard version of the elastic net problem.
}
