% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FGSPCAUtils.R
\name{FSGFunL2}
\alias{FSGFunL2}
\title{The Feature Selection and Grouping Function FSGFun with L2 loss objective.}
\usage{
FSGFunL2(
  x,
  y,
  beta,
  tau_S,
  lambda1,
  lambda2,
  lambda3,
  v = 1,
  c = 1.02,
  iter_m_max = 100,
  iter_k_max = 50,
  condition_tol = 1e-05,
  nnConstraint = FALSE,
  sparseTruncated = TRUE,
  loss_return = FALSE
)
}
\arguments{
\item{x}{the data matrix \eqn{X_{n\times p}}, where \eqn{n} is the number of observations,
\eqn{p} is the number of features.}

\item{y}{the response vector \eqn{Y_{n\times 1}} with length \eqn{n}}

\item{beta}{\eqn{\beta}, the estimation of \eqn{\beta}}

\item{tau_S}{\eqn{\tau}, a global \eqn{\tau}, which is assigned to \eqn{\tau_1=\tau_2=\tau .}}

\item{lambda1}{\eqn{\lambda_1}, the tuning parameter corresponding to \eqn{p_1(\cdot)}}

\item{lambda2}{\eqn{\lambda_2}, the tuning parameter corresponding to \eqn{p_2(\cdot)}}

\item{lambda3}{\eqn{\lambda_3}, the tuning parameter corresponding to \eqn{p_3(\cdot)}}

\item{v}{the initial value of the Lagrange multiplier, with default \code{v=1.0}.}

\item{c}{the acceleration constant to speed up the convergence procedure, default \code{c=1.02}.}

\item{iter_m_max}{the maximum number of the outer iterations (i.e. the m-iteration), default \code{iter_m_max=100}}

\item{iter_k_max}{the maximum number of the inner iterations (i.e. the k-iteration), default \code{iter_k_max=50}}

\item{condition_tol}{the conditional tolerance of both the outer and inner iterations, default \code{condition_tol=1e-5}}

\item{nnConstraint}{Boolean, indicating the non-negative constraint is true or false, default \code{FALSE}}

\item{sparseTruncated}{Boolean, indicating whether to use the truncated L1 penalty or not for sparsity,
default \code{TRUE}}

\item{loss_return}{Boolean, indicating whether return the loss or not, default \code{FALSE}}
}
\value{
the solution \eqn{\beta} or with the corresponding loss if \code{loss_return=TRUE}
}
\description{
The FSGFun with L2 loss is a subproblem of the FGSPCA with L2 loss.
}
\details{
The parameters of \code{x}, \code{y}, \code{beta}, \code{lambda1}, \code{lambda2}, \code{lambda3} are in inherit from \code{ObjFunL2}.
}
