% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FGSPCAUtils.R
\name{ObjFun_FG_S_PCA_L2}
\alias{ObjFun_FG_S_PCA_L2}
\title{Objective Function of Feature Grouping and Sparse PCA with L2 loss.}
\usage{
ObjFun_FG_S_PCA_L2(
  x,
  A,
  B,
  tau_S,
  lambda1,
  lambda2 = 0.1,
  lambda3 = 0.1,
  nnConstraint = FALSE,
  sparseTruncated = TRUE
)
}
\arguments{
\item{x}{the data matrix \eqn{X_{n\times p}}, where \eqn{n} is the number of observations,
\eqn{p} is the number of features.}

\item{A}{matrix \eqn{A} in the FGSPCA problem}

\item{B}{matrix \eqn{B} in the FGSPCA problem}

\item{tau_S}{\eqn{\tau}, a global \eqn{\tau}, which is assigned to \eqn{\tau_1=\tau_2=\tau .}}

\item{lambda1}{\eqn{\lambda_1}, the tuning parameter corresponding to \eqn{p_1(\cdot)}}

\item{lambda2}{\eqn{\lambda_2}, the tuning parameter corresponding to \eqn{p_2(\cdot)}}

\item{lambda3}{\eqn{\lambda_3}, the tuning parameter corresponding to \eqn{p_3(\cdot)}}

\item{nnConstraint}{Boolean, indicating the non-negative constraint is true or false, default \code{FALSE}}

\item{sparseTruncated}{Boolean, indicating whether to use the truncated L1 penalty or not for sparsity,
default \code{TRUE}}
}
\description{
Objective Function of Feature Grouping and Sparse PCA with L2 loss.
}
\details{
The L2 loss is defined as follows
\deqn{ (X - X B A^T)^2 .}

The penalty function with three parts is defined as follows,
\deqn{ \Psi(\beta) = \lambda_{1}p_1(\beta) + \lambda_{2}p_2(\beta) + \lambda_{3}p_3(\beta)}
where
\deqn{p_1(\beta) = \sum_{j=1}^p \min\{\frac{|\beta_j|}{\tau_1}, 1\} , }
\deqn{p_2(\beta) = \sum_{j < j', (j, j') \in E} \min\{\frac{|\beta_j - \beta_{j'}|}{\tau_2}, 1\}, }
\deqn{p_3(\beta) = \sum_{j=1}^p (\min\{\beta_j, 0\})^2 . }

With the setting of \verb{lambda2=0, nnConstraint=FALSE, sparseTruncated=FALSE},
it corresponds to the same objective function of sparse PCA; \\

\verb{ObjFun_FG_S_PCA_L2(x, A, B, tau_S, lambda1, lambda2=0, lambda3=0.1, nnConstraint=FALSE, sparseTruncated=FALSE}

With setting of \verb{lambda2=0, nnConstraint=FALSE, sparseTruncated=TRUE},
corresponds to Objective function of truncated sparse PCA. \\

\verb{ObjFun_FG_S_PCA_L2(x, A, B, tau_S, lambda1, lambda2=0, lambda3=0.1, nnConstraint=FALSE, sparseTruncated=TRUE}

By tuning the different setting of \code{lambda2}, \code{nnConstraint}, \code{sparseTruncated},
we get different combination of models.
}
