% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FGSPCAUtils.R
\name{ObjFunL2}
\alias{ObjFunL2}
\title{The objective function using L2 loss with three penalty functions}
\usage{
ObjFunL2(
  x,
  y,
  tau_S1,
  tau_S2,
  lambda1,
  lambda2,
  lambda3,
  beta,
  Bjj,
  SF,
  SFc,
  SE,
  SEc,
  SN
)
}
\arguments{
\item{x}{the data matrix \eqn{X_{n\times p}}, where \eqn{n} is the number of observations,
\eqn{p} is the number of features.}

\item{y}{the response vector \eqn{Y_{n\times 1}} with length \eqn{n}}

\item{tau_S1}{\eqn{\tau_1}, the controlling parameter corresponding to \eqn{p_1(\cdot)},
which determines when the small values of \eqn{|\beta_j|} will be penalized.}

\item{tau_S2}{\eqn{\tau_2}, the controlling parameter corresponding to \eqn{p_2(\cdot)},
which determines when the small difference values of \eqn{|\beta_j - \beta_{j'}|} will be penalized.}

\item{lambda1}{\eqn{\lambda_1}, the tuning parameter corresponding to \eqn{p_1(\cdot)}}

\item{lambda2}{\eqn{\lambda_2}, the tuning parameter corresponding to \eqn{p_2(\cdot)}}

\item{lambda3}{\eqn{\lambda_3}, the tuning parameter corresponding to \eqn{p_3(\cdot)}}

\item{beta}{\eqn{\beta}, the estimation of \eqn{\beta}}

\item{Bjj}{a matrix of \eqn{p\times p} with element
\eqn{\beta_{jj'} = \beta_{j}-\beta_{j'} } .}

\item{SF}{the \eqn{\mathcal{F}} set, \eqn{p}-length vector of indicator 0-1.
Its value is 1 if \eqn{|\beta_j| \leq \tau_1 }. Otherwise 0.}

\item{SFc}{the \eqn{\mathcal{F}^{c}} set, \eqn{p}-length vector of indicator 0-1.
Its value is 1 if  \eqn{|\beta_j| > \tau_1 }. Otherwise 0.}

\item{SE}{the \eqn{\mathcal{E}} set, a matrix of \eqn{p\times p} with indicator 0-1.
Its value is 1 if \eqn{|\beta_{j}-\beta_{j'}| \leq \tau_2}. Otherwise 0.
Note if \eqn{\tau_2=0, j=j'}, then \eqn{0<=0 } is true, the diagonal of \eqn{\mathcal{E}} is 1.
We should set \code{SEc <- (1-SE)}, \code{SE <- SE - diag(p)} after the calculation of \eqn{\mathcal{E}}.}

\item{SEc}{the \eqn{\mathcal{E}^{c}} set, a matrix of \eqn{p\times p} with indicator 0-1.}

\item{SN}{the \eqn{\mathcal{N}} set, \eqn{p}-length vector of indicator 0-1.
Its value is 1 when \eqn{\beta_j < 0}, corresponding to \eqn{\min(\beta_j, 0)}.}
}
\value{
the value of the objective function
}
\description{
The objective function using L2-loss with three penalty functions. An extension of the elastic net regression.
}
\details{
The objective function using L2 loss is defined as follows,
\deqn{ \frac{1}{2n} \|Y- X\beta\|^2 + \lambda_{1}p_1(\beta) + \lambda_{2}p_2(\beta) + \lambda_{3}p_3(\beta) .}
The three penalties are as follows,
\deqn{p_1(\beta) = \sum_{j=1}^p \min\{\frac{|\beta_j|}{\tau_1}, 1\} , }
\deqn{p_2(\beta) = \sum_{j < j', (j, j') \in E} \min\{\frac{|\beta_j - \beta_{j'}|}{\tau_2}, 1\}, }
\deqn{p_3(\beta) = \sum_{j=1}^p (\min\{\beta_j, 0\})^2 . }
}
