% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FGSPCAUtils.R
\name{PenaltyFun}
\alias{PenaltyFun}
\title{The penalty function consisting of three parts.}
\usage{
PenaltyFun(
  beta,
  tau_S,
  lambda1,
  lambda2,
  lambda3,
  nnConstraint = FALSE,
  sparseTruncated = TRUE
)
}
\arguments{
\item{beta}{\eqn{\beta}, the estimation of \eqn{\beta}}

\item{tau_S}{\eqn{\tau} a global \eqn{\tau}, which is assigned to \eqn{\tau_1=\tau_2=\tau .}}

\item{lambda1}{\eqn{\lambda_1}, the tuning parameter corresponding to \eqn{p_1(\cdot)}}

\item{lambda2}{\eqn{\lambda_2}, the tuning parameter corresponding to \eqn{p_2(\cdot)}}

\item{lambda3}{\eqn{\lambda_3}, the tuning parameter corresponding to \eqn{p_3(\cdot)}}

\item{nnConstraint}{Boolean, indicating the non-negative constraint is true or false,
default \code{FALSE}}

\item{sparseTruncated}{Boolean, indicating whether use the truncated L1 penalty or not for sparsity,
default \code{TRUE}}
}
\description{
The penalty function consisting of three parts.
}
\details{
The three penalties are as follows,
\deqn{p_1(\beta) = \sum_{j=1}^p \min\{\frac{|\beta_j|}{\tau_1}, 1\} , }
\deqn{p_2(\beta) = \sum_{j < j', (j, j') \in E} \min\{\frac{|\beta_j - \beta_{j'}|}{\tau_2}, 1\}, }
\deqn{p_3(\beta) = \sum_{j=1}^p (\min\{\beta_j, 0\})^2 . }
}
