% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FGSPCAUtils.R
\name{rootmatrix}
\alias{rootmatrix}
\title{Calculate the root matrix of the given square matrix using eigendecomposition}
\usage{
rootmatrix(x)
}
\arguments{
\item{x}{the given square matrix with size \eqn{n\times n}}
}
\value{
the root matrix
}
\description{
Calculate the root matrix of the given square matrix using eigendecomposition
}
