% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FGSPCAUtils.R
\name{solvebeta}
\alias{solvebeta}
\title{The SPCA solver of beta for the elastic net problem}
\usage{
solvebeta(
  x,
  y,
  paras,
  max.steps,
  sparse = "penalty",
  eps = .Machine$double.eps
)
}
\arguments{
\item{x}{the data matrix \eqn{X_{n\times p}}}

\item{y}{the response vector \eqn{Y_{n\times 1}}}

\item{paras}{the combination of parameters \eqn{(\lambda_2, \lambda_1)}\\
\code{lambda = paras[1]} (which is lambda2) and \code{lambda1 = paras[2]} .}

\item{max.steps}{\code{100} (default) the maximum number of steps for the updating}

\item{sparse}{\code{sparse = c("penalty", "varnum")}}

\item{eps}{the tolerance of the stopping criterion for the termination}
}
\value{
the solution \eqn{\beta} in each subproblem
}
\description{
The beta solver using the sparse PCA.
}
\details{
The standard objective function of elastic net is
\deqn{1/(2n) \| Y-X\beta\|_2^2 + \lambda (\alpha \|\beta\|_1 + (1-\alpha )/2\|\beta\|_2^2) .}
But here we use the following objective function
\deqn{1/(2n) \| Y-X\beta\|_2^2 + \lambda_1 \|\beta\|_1 + \lambda_2/2 \|\beta\|_2^2 .}
}
\references{
\itemize{
\item Zou, Hui, Trevor Hastie, and Robert Tibshirani. "Sparse principal component analysis."
Journal of computational and graphical statistics 15.2 (2006): 265-286.
}
}
