% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FGSPCAUtils.R
\name{total_variance_explained}
\alias{total_variance_explained}
\title{Adjusted Total Variance by Zou. et.al.(2006)}
\usage{
total_variance_explained(x, NB)
}
\arguments{
\item{x}{the data matrix by \eqn{n\times p}. It should not be the sample covariance matrix,
nor the correlation matrix.}

\item{NB}{normalized matrix of loading vectors}
}
\value{
a list of \code{c(pev_svd, pev_tr, pev_spca, cum_svd, cum_tr, cum_spca)}.
}
\description{
To calculate the Total Variance Explained by the new normalized loadings NB.
The loadings of the ordinary principal components are uncorrelated and orthogonal,
where the orthogonal property, and the uncorrelated property are satisfied together.
}
\references{
\itemize{
\item Zou, Hui, Trevor Hastie, and Robert Tibshirani. "Sparse principal component analysis."
Journal of computational and graphical statistics 15.2 (2006): 265-286.
}
}
